src/models/ercot/exp0/model_adapters/tree_models.py
"""
Tree-based Model Adapters for ERCOT DART Prediction

This module provides adapters for tree-based models including:
- Random Forest
- Gradient Boosting
- XGBoost

All models follow the same adapter interface for consistent usage.
"""

import pickle
from pathlib import Path
from typing import Any
from typing import Dict
from typing import Optional
from typing import Union

import numpy as np
import pandas as pd
import torch
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error
from sklearn.metrics import mean_squared_error
from sklearn.metrics import r2_score

from src.models.ercot.exp0.model_adapters.base_adapter import BaseModelAdapter


class TreeModelAdapter(BaseModelAdapter):
    """
    Adapter for tree-based models.

    Note: Implementation will be added in Phase 2 of the modeling approach.
    This is a placeholder to establish the architecture.
    """

    def __init__(self, model_type: str = "random_forest", **model_kwargs):
        super().__init__(
            f"tree_{model_type}", {"model_type": model_type, **model_kwargs}
        )

        # TODO: Implement in Phase 2
        raise NotImplementedError("Tree models will be implemented in Phase 2")

    def fit(
        self,
        X: Union[np.ndarray, torch.Tensor, pd.DataFrame],
        y: Union[np.ndarray, torch.Tensor, pd.Series],
        X_val: Optional[Union[np.ndarray, torch.Tensor, pd.DataFrame]] = None,
        y_val: Optional[Union[np.ndarray, torch.Tensor, pd.Series]] = None,
    ) -> Dict[str, Any]:
        # TODO: Implement tree model training
        raise NotImplementedError("Tree models will be implemented in Phase 2")

    def predict(
        self, X: Union[np.ndarray, torch.Tensor, pd.DataFrame]
    ) -> Union[np.ndarray, torch.Tensor]:
        # TODO: Implement tree model prediction
        raise NotImplementedError("Tree models will be implemented in Phase 2")

    def save_model(self, filepath: Union[str, Path]) -> None:
        # TODO: Implement model saving
        raise NotImplementedError("Tree models will be implemented in Phase 2")

    def load_model(self, filepath: Union[str, Path]) -> None:
        # TODO: Implement model loading
        raise NotImplementedError("Tree models will be implemented in Phase 2")
