pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ercot_dart"
version = "0.1.0"
description = "ERCOT DART price prediction system"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "numpy",
    "scikit-learn",
    "plotly",
    "requests",
    "python-dotenv",
    "tqdm",
    "torch",
    "pathlib",
]

[project.optional-dependencies]
dev = [
    "flake8",
    "isort",
    "black",
    "flake8-import-order",
    "pytest",
]

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
known_first_party = ["src"]
force_single_line = true
line_length = 88

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "build",
    "dist"
]
# Import order enforcement
import-order-style = "pycharm"
application-import-names = "src" 